Update FUNDING.yml
By moving the site to Firebase to let anyone upload icons, regardless of their GitHub knowledge, some costs are involved. This won't be much, all I hope is to cover the server costs. I'll be as transparent as I can and will be posting the billing costs at the end of each month in here